Estimation

Model setup

The first step of the estimation routine is for the user to define her own model type through:

MomentMatching.EstimationSetup — Type

struct EstimationSetup{U<:EstimationMode}

Description

Structure to store setup of matching moments estimation procedure.

Fields

mode: Estimation mode.

modelname: Submethod for estimation. String encoding which parameters and how are estimated.

typemom: Submethod for estimation. String encoding which moments are targeted.

For instance, running:

defines an EstimationSetup type with mode called Model1, modelname called benchmark and typemom set1.

Once EstimationSetup has been defined,

MomentMatching.AuxiliaryParameters — Type

abstract type AuxiliaryParameters

Description

Supplies non-estimated auxuliary parameters, which are needed to compute model moments for any guess of the estimated parameters. Assumed to be non-random and constant over the estimation procedure.

MomentMatching.PredrawnShocks — Type

abstract type PredrawnShocks

Description

Supplies pre-drawn shocks, which are needed to compute model moments for any guess of the estimated parameters. Assumed to be random, but constant over the estimation procedure (so that the minimization procedure works well). They are however re-simulated when performing bootstrap.

MomentMatching.PreallocatedContainers — Type

abstract type PreallocatedContainers

Description

Supplies empty arrays for computations to calculate model moments as few memory allocations as possible.

%@docs %NumParMM %

MomentMatching.ParMM — Type

struct ParMM{S<:AbstractFloat}

Description

Structure to store matching moments estimation inputs.

Fields

lb_global: Lower bound for parameters in global stage

ub_global: Upper bound for parameters in global stage

lb_hard: Hard lower bound for parameters - being outside brings penalty in local stage

ub_hard: Hard upper bound for parameters - being outside brings penalty in local stage

labels: Labels of estimated parameters

momdat: Moments from the data to match

mmomdat: Normalizing factor moments from the data to match

W: Weighting matrix

mdifrec: Predetermined quantity to re-center moment condition

Numerical routines

The main function to run the estimation routine is:

MomentMatching.estimation — Function

estimation(estset::EstimationSetup;<keyword arguments>)

Estimate model parameters given instance of EstimationSetup.

Can be customized if non-default estimation cases have to be performed. Accepts initial guess(es) when only local stage is performed.

Keyword arguments

npmm::NumParMM=NumParMM(estset): numerical parameters for estimation.

cs::ComputationSettings=ComputationSettings(): computational options.

aux::AuxiliaryParameters=AuxiliaryParameters(estset): any additional pre-determined (i.e., not to be estimated) model-specific parameters.

presh::PredrawnShocks=PredrawnShocks(estset, aux): pre-generated random shocks.

xlocstart=[[1.0]]::Vector{Vector{Float64}}: starting point for local optimization when only local stage is performed.

saving::Bool=true: whether to save the final estimation results.

saving_bestmodel::Bool=saving: whether to save the results for the best parameters vectors.

number_bestmodel::Integer=1: for how many best parameter vectors to save results.

filename_suffix::String="": suffix to use in output filenames.

errorcatching::Bool=false: if true the estimation routine does not stop if an error happens when computing moments for a given parameter vector guess.

vararg...: additional arguments passed.

In addition to the model setup elements described above, the user can also define her preferred computational settings:

MomentMatching.ComputationSettings — Type

struct ComputationSettings{T<:Integer}

Description

Computational settings for performing parallel computing either locally or on a cluster.

Fields

location: Where computation is performed. 'local' and 'slurm' are supported presently Default: local

num_procs: Number of processes used by estimation and bootstrapping routines. Giving 1 avoids adding any new process. On Slurm cluster this has to be equal to :nodes * :ntaskspernode Default: 1

num_tasks: Number of tasks spawned per process in estimation and bootstrapping routines. Giving somewhat more than the number of actual threads is probably a good idea to avoid idleness. Setting it equal to one implies that only one thread per process will be used by functions in this package (even if more are present). Note that in this case multithreading within the internal function is possible as long that is written in a data-race free way. Default: Threads.nthreads() * 2

num_threads: Number of threads to be used when starting each new process. Relevant only with multiprocessing. On Slurm cluster this has to be equal to :cpuspertask Default: Int(floor(Threads.nthreads() / num_procs))

maxmem: Trigger intensive garbage collection at this memory usage Default: -1

clustermanager_settings: Other settings Default: Dict(:x => "")

Furthermore, the user can decide:

whether to save the final estimation results through saving;

whether to save the evaluated moments for the best (i.e., those returning the lowest objective function values) number_bestmodel points through saving_bestmodel;

whether to add a specific suffix to saved files via filename_suffix;

whether to proceed with the estimation routine until the end if evaluation of the objective function at a specific candidate point returns an error errorcatching.

Additional model-specific arguments needed for model initialization can be passed through via vararg....

Results of the estimation routine are stored in a dedicated structure:

MomentMatching.EstimationResult — Type

struct EstimationResult{S<:AbstractFloat, T<:Integer, U<:AuxiliaryParameters, V<:PredrawnShocks}

Description

Structure to store output of matching moments estimation procedure.

Fields

npmm: Numerical parameters

aux: Auxiliary inputs

presh: Predrawn shocks

xlocstart: Starting parameters - relevant when only local stage is performed

pmm: Moment estimation inputs

fglo: Objective function value in global stage, sorted in increasing order

xglo: Parameter combinations checked in global stage, sorted according to objective function value

momglo: Moments from model, global stage

floc: Objective function value in local stage, sorted in increasing order

xloc: Parameter combinations checked in local stage, sorted according to objective function value

momloc: Moments from model, local stage

conv: Logical, if true convergence criterion at minimum is satisfied in the local stage

Global stage

Local stage

Comparison with other algorithms

Multithreading and multiprocessing

Locally

Cluster

Two-step estimation